9-serverless/pyproject.toml
[project]
name = "9-serverless"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.4.0",
    "pillow>=12.0.0",
    "requests>=2.32.5",
    "torchvision>=0.24.1",
]
